Add windows support
Description
===========

This patch aims to add windows support for the jni bindings of the unity
version manager. The main change is the update of `uvm_core` to version
`0.1.1` which includes major changes when it comes to the internal
install logic. To be able to use __uvm__ on windows we need a custom
installer routine tailored for windows. The latest version of `uvm_core`
includes just that plus addtional fixes and changes.

* usage of error_chain crate
* custom logging

Build Setup
-----------

This patch also introduces a new custom `Jenkinsfile` to build the
native libraries on both macOS and windows. The final library jar
published to bintray includes both osx and windows native libraries.

The gradle rust plugin has some issues on windows so we use the
preinstalled rust setup on windows.

Changes
=======

* ![UPDATE] `uvm_core` to `0.1.1`
* ![IMPROVE] install.rs to support windows
* ![IMPROVE] build and publish workflow