Improve Installation.atLocation
Description
===========

I made a mistake when implementing a fix
https://github.com/wooga/atlas-unity/pull/52
The plugin tries to retrieve the version of Unity being executed to
determine the arguments to pass for the `logFile` option.
The fixed implementation is passing the path to the unity executable
because the plugin isn't aware of a difference. So I decided to fix this
issue in this library instead of the plugin.

I could or should push this logic in some form to the `uvm_core` library
but won't do it at the moment. It is quite hard to find a nice looking
implementation for this issue. The current solution works but is not
very elegant.

The proposed implementation adjusts the installation path provided
before passing it down to `uvm`. It checks if the path points to a file
instead of a directory and returns a new path a few levels higher. This
is a very brute force way of solving this issue for now and lacks error
handling or transparency. In a worst case the resulting error from `uvm`
could confuse users of this library when they pass a file pointing to a
file and get reports of invalid unity location on a different path.

Changes
=======

![IMPROVE] `Installation.atLocation` fetch installations with executable path