Bootstrap preinstalled Unity version in tests (#64)
Description
===========

Instead of using a unityversion being installed on the test system this
patch ensures a unity installation at test runtime. It uses a spock
extension library [spock-unity-version-manager-extension]. The irony is
that this extension uses this library under the hood. Classic Henn and
Egg problem. I don't see a huge issue with that.

Changes
=======

* ![IMPROVE] bootstrap preinstalled Unity versions in tests

[spock-unity-version-manager-extension]: https://github.com/wooga/spock-unity-version-manager-extension