Fix NativeLoader on windows
Description
===========

The `NativeLoader` has a small but on windows
when loading libraries by name when the ending is
already provided. eg uvm_jni.dll. For all other platforms
the loader checks if the extension is already provided before
attempting to add it again. For windows this check was missing.

Changes
========

* ![FIX] ![WIN] `NativeLoader` on windows.