Increase unity version support
Description
===========

With this change the JNI library will gain more official
support for the newer Unity modules. Unity 2020 and 2021
worked for the most common modules but had no support for the
newly added ones (LinuxIL2CPP, LinuxServer, etc)
This patch updates the main dependency `uvm-install2` to
version `0.5.0` which adds all changes needed to install
the new modules.
The new version also installs the android ndk/sdk versions
in the correct way for Unity 2020 and beyond.

I updated the java enum `net.wooga.uvm.Component` and adjusted
the mapping on the rust side.

Changes
=======

* ![UPDATE] `uvm-install2` to `0.5.0`
* ![IMPROVE] version support for modules from 2020-2022