Improve jni library setup and logging
Descritption
============

This is a test to see if I can improve the logging
setup of the JNI library. I would also like to adjust
the rust sources and mixxed this into this one patch.
The main problem is the caching of the class references which I
need to do in an `unsafe` manner. The idea is to have a
`jni_cache` module which loads the method/class/field references at
loadtime of the shared libary. This should improve the performance.

It also sets up a custom log writer which sends all messages to
a logger on the Java side. I need to measure what this would mean in
real world usage since the bottleneck here is JNI. I allready cache
the created logger obj as a global reference, Means it should not get
garbage collected between JNI calls.

Changes
=======

* ![IMPROVE] crate layout
* ![ADD] `jni` global method/class cache
* ![ADD] custom log writer to send logs from rust to java