Fix iOSSupport module installation destination on linux (#114)
Description
===========

The `iOSSupport` module is the only package on `macOS` which is archived
based on a different base directory than all other packages. This is sadly
also true for the linux version of the module. So I adjusted the installation
destination when the package destination contains `iOSSupport` in the path.

This patch is done through an update of `uvm_install_core` to version `0.5.1`
which is only reflected in the `Cargo.lock` file.

Changes
=======

![FIX]![LINUX] `iOSSupport` module installation destination