Build linux variant with debian stretch image
Description
===========

I tried to fix a glibc issue when running the compiled
jni library with ubuntu 18.4. I thought 18.4 is based on
buster (which I guess it is) but buster is using GLIBC 2.28
and ubutu 18.4 supports up to 2.27.
I want to test out if using stretch is possible for this
specific build and later try to find a better working solution.

Changes
=======

* ![FIX] ![LINUX] build with Debian Stretch image