Add aarch64 apple support
Description
===========

With the introduction of the AppleM1 and native java distributions it is
needed to add support the `aarch64` target to the dylib. Apple supports the
nothion of universal libries which contain multiple architectures. Rust does
not support to build these kinds of libraries directly from cargo.
But the basic steps to generate these universal libraries are simple.

1. build the lib in multiple architecture
2. call the `lipo` cli tool to generate a universal library from the provided lib.

In order to be able to crosscompile to both architectures we need the help of
`rustup` to install the needed targets. The new plugin `net.wooga.rustup` takes
care of this and we hook the installed cargo and toolchain into the `net.wooga.rust`
plugin. I added some basic magic tasks that compile the lib twice and combine it to
one library.

Changes
=======

* ![ADD] support for AppleM1 `aarch64`