Improve detect unity project version and install API (#119)
Description
===========

The latest updated to `uvm-install2` crate brought the ability to
fetch the revision hash from the `ProjectVersion.txt` file. This means
that even if the webservice, which provides the mapping from unity version string
to version revision hash, is outdated or doesn't know the *maybe* valid version
requested it is still able to download all installer resources. This can become crucial
for custom internal beta releases from Untiy which are not tracked in the offical release
files and pages.

This change consists of a small API change in `UnityVersionManager` and a verification
test change in the tests to verify that it is possible to request a unity editor installation
with a version containing a revision hash.

> Note:
> Unity version is a complicated mess. In praxis most humans express the version in the form of
> `2020.3.20f1`. But sadly that is not enough to install a unity version. All installation resources
> need a second information which is the *revision* (term is not quite clear to me but that is the latest term
> I saw in the new graphQL endpoint). Together with the version and the revision on can find the installer
> manifest which includes all the informations where installers are located etc.
> UVM stores unity versions not as a string object like we do in java land. For one the version needs to be
> able to be sorted etc. And the revision hash is treated as an optional part which can be requested. If the version
> doesn't know the revision it will try to locate it from the versions service or local cache.
> This is the key here as the latest patches allowed to parse a version from a string with the revision included!

The mentioned patch in `UnityVersionManager` updates the API for `detectProjectVersion` method.
I added a new parameter `detectProjectVersion(File, boolean)`. I also added a backwards compatible API which
will call the method with the second paramter set to `false` which results in the old behavior.

With the new second parameter set to `true`, the underlying library will return the unity version printed in the form of
`2020.3.20f1 (REVSION)`. This version can be directly used to request the installation. The process is transparent. I
decided to add/update the API rather than just changeing the return value because I want to be able to control what format
the version will be returned in. In the future I think it would be better to create a custom version object which
maps better to the one in rust. This would also allow to sort and compare versions.

Changes
=======

* ![IMPROVE] unity project version detection API