Add additional flag to opt out of sync module installation
Description
===========

This patch adds a API addition which is fully backwards compatible
both in API and behavior. I added a new `installUnityEditor`
overload which adds a new property `syncChildComponents`.
I used the naming scheme from uvm. So it's component instead of
modules. The current default behavior is to install all sub modules
of a given module. e.g. Android will also install android NDK, SDK etc.

We want to opt out of this. For this library I would like to keep
the old behavior and fix it on the calling side.

Changes
=======

* ![ADD] boolean flag to opt out of sync module installation